+++ packages/components/Badge/src/Badge.tsx
@@ -19,11 +19,13 @@ export const badgeLookup = (layer: string, userProps: BadgeProps): boolean => {
   return (
     userProps[layer] ||
     layer === userProps['appearance'] ||
+    (!userProps['appearance'] && layer === 'filled') ||
     layer === userProps['size'] ||
-    (!userProps['size'] && layer === 'large') ||
+    (!userProps['size'] && layer === 'medium') ||
     layer === userProps['shape'] ||
     (!userProps['shape'] && layer === 'circular') ||
     layer === userProps['badgeColor'] ||
+    (!userProps['badgeColor'] && layer === 'brand') ||
     (I18nManager.isRTL && layer === 'rtl')
   );
 };