+++ packages/framework/memo-cache/src/getCacheEntry.ts
@@ -22,6 +22,9 @@ export type CacheEntry<T, TGet = any> = {
  * @param key - which key of that entry to ensure the value for
  */
 function ensureAndReturn(entry: CacheEntry<any>, key: keyof CacheEntry<any>): CacheEntry<any> | { [key: string]: CacheEntry<any> } {
+  if ((key as string) === '__proto__' || (key as string) === 'constructor' || (key as string) === 'prototype') {
+    throw new Error('Invalid key');
+  }
   return (entry[key] = entry[key] || {});
 }