+++ apps/win32/CHANGELOG.md
@@ -1,8 +1,32 @@
 # Change Log - @fluentui-react-native/tester-win32
 
-This log was last generated on Mon, 15 Apr 2024 19:08:10 GMT and should not be manually modified.
+<!-- This log was last generated on Wed, 11 Jun 2025 19:34:43 GMT and should not be manually modified. -->
 
 <!-- Start content -->
+
+## 0.38.45
+
+Wed, 11 Jun 2025 19:34:43 GMT
+
+### Patches
+
+- Bump rex as a test of publishing (ruaraki@microsoft.com)
+
+## 0.38.43
+
+Mon, 10 Mar 2025 20:26:46 GMT
+
+### Patches
+
+- Updating some rnx=kit packages (ruaraki@microsoft.com)
+
+## 0.38.26
+
+Fri, 27 Sep 2024 05:53:19 GMT
+
+### Patches
+
+- Callout test page additions to test new focusWindow/blurWindow APIs (ppatboyd@outlook.com)
 
 ## 0.38.2