+++ packages/codemods/CHANGELOG.md
@@ -1,8 +1,24 @@
 # Change Log - @fluentui-react-native/codemods
 
-This log was last generated on Fri, 15 Mar 2024 16:26:18 GMT and should not be manually modified.
+<!-- This log was last generated on Wed, 29 Jan 2025 23:43:54 GMT and should not be manually modified. -->
 
 <!-- Start content -->
+
+## 0.5.14
+
+Wed, 29 Jan 2025 23:43:54 GMT
+
+### Patches
+
+- fix tsconfig for jest types (sanajmi@microsoft.com)
+
+## 0.5.13
+
+Mon, 23 Sep 2024 18:26:22 GMT
+
+### Patches
+
+- Updating jscodeshift (ruaraki@microsoft.com)
 
 ## 0.5.11