+++ packages/components/Badge/CHANGELOG.md
@@ -1,8 +1,24 @@
 # Change Log - @fluentui-react-native/badge
 
-This log was last generated on Fri, 12 Apr 2024 21:02:37 GMT and should not be manually modified.
+<!-- This log was last generated on Wed, 29 Jan 2025 23:43:56 GMT and should not be manually modified. -->
 
 <!-- Start content -->
+
+## 0.11.2
+
+Wed, 29 Jan 2025 23:43:56 GMT
+
+### Patches
+
+- remove enzyme tests (sanajmi@microsoft.com)
+
+## 0.11.0
+
+Thu, 16 Jan 2025 19:12:11 GMT
+
+### Minor changes
+
+- Changing size default and fixing appearance default for Badge (ruaraki@microsoft.com)
 
 ## 0.9.1