+++ packages/deprecated/theming-ramp/CHANGELOG.md
@@ -1,8 +1,16 @@
 # Change Log - @uifabricshared/theming-ramp
 
-This log was last generated on Thu, 11 Apr 2024 18:08:41 GMT and should not be manually modified.
+<!-- This log was last generated on Wed, 29 Jan 2025 23:43:53 GMT and should not be manually modified. -->
 
 <!-- Start content -->
+
+## 0.20.2
+
+Wed, 29 Jan 2025 23:43:53 GMT
+
+### Patches
+
+- fix tsconfig for jest types (sanajmi@microsoft.com)
 
 ## 0.20.0