+++ packages/framework/composition/CHANGELOG.md
@@ -1,8 +1,16 @@
 # Change Log - @fluentui-react-native/composition
 
-This log was last generated on Thu, 11 Apr 2024 18:08:42 GMT and should not be manually modified.
+<!-- This log was last generated on Wed, 29 Jan 2025 23:43:54 GMT and should not be manually modified. -->
 
 <!-- Start content -->
+
+## 0.11.2
+
+Wed, 29 Jan 2025 23:43:54 GMT
+
+### Patches
+
+- fix tsconfig for jest types (sanajmi@microsoft.com)
 
 ## 0.11.0