+++ packages/theming/apple-theme/CHANGELOG.md
@@ -1,8 +1,16 @@
 # Change Log - @fluentui-react-native/apple-theme
 
-This log was last generated on Thu, 11 Apr 2024 18:08:41 GMT and should not be manually modified.
+<!-- This log was last generated on Wed, 29 Jan 2025 23:43:55 GMT and should not be manually modified. -->
 
 <!-- Start content -->
+
+## 0.28.2
+
+Wed, 29 Jan 2025 23:43:55 GMT
+
+### Patches
+
+- fix tsconfig for jest types (sanajmi@microsoft.com)
 
 ## 0.28.0