+++ packages/utils/interactive-hooks/CHANGELOG.md
@@ -1,8 +1,24 @@
 # Change Log - @fluentui-react-native/interactive-hooks
 
-This log was last generated on Thu, 11 Apr 2024 18:08:38 GMT and should not be manually modified.
+<!-- This log was last generated on Wed, 29 Jan 2025 23:43:53 GMT and should not be manually modified. -->
 
 <!-- Start content -->
+
+## 0.27.4
+
+Wed, 29 Jan 2025 23:43:53 GMT
+
+### Patches
+
+- remove enzyme tests (sanajmi@microsoft.com)
+
+## 0.27.2
+
+Thu, 16 Jan 2025 19:12:10 GMT
+
+### Patches
+
+- Fix controlled toggling (ruaraki@microsoft.com)
 
 ## 0.27.0